move validate_ert_design_matrix to separate module
